fixed bug that prevented the user from being able to log in immediately
