+++ src/components/ApplicationViews.js
@@ -2,7 +2,6 @@ import React, { Component } from 'react';
 import { Route } from "react-router-dom";
 import RepresentativeManager from "../modules/RepresentativeManager"
 import ProfileManager from "../modules/ProfileManager"
-import LoginManager from "../modules/LoginManager"
 import LocalList from "./local/LocalList"
 import ProfileList from "./profile/ProfileList"
 import Login from "./login/Login"
@@ -24,21 +23,7 @@ export default class ApplicationViews extends Component {
     notes: [],
   }
 
-  // Check the user input against the database and then, if found, set the state as the found user
-  authenticateUser = (userInput, userPass) => {
-    LoginManager.findUser(userInput, userPass)
-      .then(foundUser => {
-          this.setState({
-            users: foundUser
-          })
-        })
-  }
-
   componentDidMount() {
-    // get the current session user
-    let sessionUser = sessionStorage.getItem("user");
-    let sessionUserId = Number(sessionUser);
-
     // currently hardcoding the address until the search functionality is implemented
     let address = "405%20North%20Jefferson%20St.%20Winchester%20Tennessee"
 
@@ -52,14 +37,6 @@ export default class ApplicationViews extends Component {
         })
     })
 
-  // get all the politicians that the active user has saved
-    ProfileManager.getAllUserPoliticians(sessionUserId)
-    .then(allUsersSavedPoliticians => {
-        this.setState({
-          savedPoliticians: allUsersSavedPoliticians.savedPoliticians
-        })
-    })
-
     ProfileManager.getAllNotes()
       .then(allNotes => {
         this.setState({
@@ -70,6 +47,8 @@ export default class ApplicationViews extends Component {
   }
 
   render() {
+    console.log(sessionStorage.getItem("user"))
+
 
     return(
       <React.Fragment>
@@ -80,7 +59,7 @@ export default class ApplicationViews extends Component {
           return <LocalList data={this.state} />
         }} />
         <Route exact path="/profile/" render={props =>{
-          return <ProfileList allData={this.state} allUserPoliticians={this.state.savedPoliticians} />
+          return <ProfileList allData={this.state} />
         }} />
         <Route exact path="/federal" render={props =>{
           return <FederalList allData={this.state} />
