+++ src/components/profile/notes/Notes.js
@@ -4,6 +4,7 @@ export default class Notes extends Component {
 
     render() {
         return (
+            // this feature with map through all of the notes for a specified politician and add them to the user's profile
             this.props.allNotes.map(note =>
                     <h5 key={note.id}>{note.content}</h5>
                 )
